Add files via upload
This is my last assignment in Business Data Management Course. In the final project, we were given case study MyTShirt that required us to create a database for MyTShirt to store detail information of customer, orders, products and modification. Later, we were required to come up with 2 possible business question and we were asked to write SQL queries to solve them and visualize them.